requirements.txt
Django>=4.2.0
django-tailwind>=3.8.0
django-background-tasks>=1.2.8
playwright>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0